refactor: extract session tracker to separate module
